# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "MechanicShop API"
  description: "This is an API to manage a Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /mechanics:
    get:
      tags:
        - Mechanics
      summary: "Get Mechanics by Pagination"
      description: "Retrieve a list of mechanics with optional pagination."
      parameters:
        - name: "page"
          in: "query"
          type: "integer"
          description: "Page number"
        - name: "per_page"
          in: "query"
          type: "integer"
          description: "Number of mechanics per page"
      responses:
        200:
          description: "List of mechanics"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Mechanic"
    post:
      tags:
        - Mechanics
      summary: "Create Mechanic"
      description: "Create a new mechanic."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic details"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/Mechanic"
        400:
          description: "Duplicate mechanic error"
          schema:
            type: object
            properties:
              message:
                type: string
                example: "A mechanic with this name or email already exists."
...
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Mechanic Login"
      description: "Allows a mechanic to log in using their email and name."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic login credentials"
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
                description: "The mechanic's email."
                example: "mike.johnson@example.com"
              name:
                type: string
                description: "The mechanic's name."
                example: "Mike Johnson"
      responses:
        200:
          description: "Login successful"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Successfully Logged In"
              auth_token:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        400:
          description: "Invalid payload"
        404:
          description: "Mechanic not found"
        401:
          description: "Invalid email or name"
  /mechanics/remove-ticket/{ticket_id}:
    delete:
      tags:
        - Mechanics
      summary: "Remove Ticket"
      description: "Allows a mechanic to remove a service ticket assigned to them."
      parameters:
        - name: "ticket_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the service ticket to remove"
      responses:
        200:
          description: "Ticket removed successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Ticket with ID 123 successfully deleted"
        404:
          description: "Ticket not found or unauthorized"
        401:
          description: "Unauthorized access"
  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: "Get Mechanic by ID"
      description: "Retrieve a mechanic's details by their ID."
      parameters:
        - name: "mechanic_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the mechanic"
      responses:
        200:
          description: "Mechanic details"
          schema:
            $ref: "#/definitions/Mechanic"
        404:
          description: "Mechanic not found"

    delete:
      tags:
        - Mechanics
      summary: "Delete Mechanic"
      description: "Delete a mechanic by their ID."
      parameters:
        - name: "mechanic_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the mechanic to delete"
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Mechanic deleted successfully"
        404:
          description: "Mechanic not found"

    put:
      tags:
        - Mechanics
      summary: "Update Mechanic"
      description: "Update a mechanic's details by their ID."
      parameters:
        - name: "mechanic_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the mechanic to update"
        - in: "body"
          name: "body"
          description: "Updated mechanic details"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: "The mechanic's name."
                example: "Mike Johnson"
              salary:
                type: number
                description: "The mechanic's salary."
                example: 60000.00
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/Mechanic"
        400:
          description: "Validation error"
        404:
          description: "Mechanic not found"
...
  /mechanics/search:
    get:
      tags:
        - Mechanics
      summary: "Search Mechanics"
      description: "Search for mechanics by name or salary filters."
      parameters:
        - name: "name"
          in: "query"
          type: "string"
          description: "Name of the mechanic"
        - name: "salary"
          in: "query"
          type: "number"
          description: "Salary filter"
        - name: "below"
          in: "query"
          type: "boolean"
          description: "Filter for salaries below the specified amount"
      responses:
        200:
          description: "List of mechanics matching the search criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Mechanic"
definitions:
...
  Mechanic:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      salary:
        type: "number"
    example:
      id: 2
      name: "Mike Johnson"
      salary: 55000.00
  CreateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      salary:
        type: "number"
    required:
      - name
      - salary
    example:
      name: "Sarah Connor"
      salary: 60000.00
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...